requirements.txt
python-telegram-bot==20.7
langchain==0.2.0
openai
chromadb
python-dotenv
pdfplumber
langchain-community
tiktoken